# docs/openapi/season-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Season Service API
  description: API documentation for ACM Season Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/workspace/tasks/{id}:
    get:
      tags:
      - task-workspace-controller
      summary: Get task
      description: Get task details by ID
      operationId: getTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    put:
      tags:
      - task-workspace-controller
      summary: Update task
      description: Update task details
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    delete:
      tags:
      - task-workspace-controller
      summary: Delete task
      description: Permanently delete task (dev only)
      operationId: deleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/tasks/{id}:
    get:
      tags:
      - task-controller
      summary: Get task detail
      description: Get task detail if its season belongs to current farmer
      operationId: getTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    put:
      tags:
      - task-controller
      summary: Update task
      description: Update task details while task and season are still editable
      operationId: updateTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    delete:
      tags:
      - task-controller
      summary: Delete task
      description: Delete task if allowed by status and season constraints
      operationId: deleteTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/workspace/tasks:
    get:
      tags:
      - task-workspace-controller
      summary: List tasks
      description: List all tasks for current user with filters
      operationId: listTasks
      parameters:
      - name: status
        in: query
        description: Filter by status
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - DONE
          - OVERDUE
          - CANCELLED
      - name: seasonId
        in: query
        description: Filter by season ID
        required: false
        schema:
          type: integer
          format: int32
      - name: q
        in: query
        description: Search by title (min 2 chars)
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sortBy
        in: query
        description: Sort by field
        required: false
        schema:
          type: string
          default: createdAt
      - name: sortDirection
        in: query
        description: Sort direction
        required: false
        schema:
          type: string
          default: desc
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
    post:
      tags:
      - task-workspace-controller
      summary: Create task
      description: Create a new task with optional season link
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
...
  /api/v1/seasons/{seasonId}/tasks:
    get:
      tags:
      - task-controller
      summary: List tasks of a season
      description: List tasks for a given season of the current farmer
      operationId: listTasks_1
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: From date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: To date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        description: Page index (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
    post:
      tags:
      - task-controller
      summary: Create task in season
      description: Create a new task within given season for current farmer
      operationId: createTask_1
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
...
  /api/v1/tasks/{id}/status:
    patch:
      tags:
      - task-controller
      summary: Update task status
      description: Update task status with validation of allowed transitions
      operationId: updateTaskStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStatusRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "404":
          description: Task not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
...
components:
  schemas:
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 4000
          minLength: 0
        seasonId:
          type: integer
          format: int32
        plannedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        notes:
          type: string
          maxLength: 4000
          minLength: 0
        assigneeUserId:
          type: integer
          format: int64
      required:
      - title
    ApiResponseTaskResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/TaskResponse"
        data:
          $ref: "#/components/schemas/TaskResponse"
        generatedAt:
          type: string
    TaskResponse:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        status:
          type: string
        plannedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        actualStartDate:
          type: string
          format: date
        actualEndDate:
          type: string
          format: date
        notes:
          type: string
        seasonId:
          type: integer
          format: int32
        seasonName:
          type: string
        userId:
          type: integer
          format: int64
        userName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 4000
          minLength: 0
        seasonId:
          type: integer
          format: int32
        plannedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        notes:
          type: string
          maxLength: 4000
          minLength: 0
        assigneeUserId:
          type: integer
          format: int64
      required:
      - title
...
    UpdateTaskStatusRequest:
      type: object
      properties:
        status:
          type: string
        actualStartDate:
          type: string
          format: date
        actualEndDate:
          type: string
          format: date
        notes:
          type: string
      required:
      - status
...
    ApiResponsePageResponseTaskResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/PageResponseTaskResponse"
        data:
          $ref: "#/components/schemas/PageResponseTaskResponse"
        generatedAt:
          type: string
    PageResponseTaskResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: object
        data:
          type: object
        generatedAt:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT